+++ spec/rspec/core/rake_task_spec.rb
@@ -221,37 +221,6 @@ module RSpec::Core
       end
     end
 
-    context "with paths including symlinked directories" do
-      include_context "isolated directory"
-      let(:project_dir) { Dir.getwd }
-
-      it "finds the files" do
-        foos_dir = File.join(project_dir, "spec/foos")
-        FileUtils.mkdir_p foos_dir
-        FileUtils.touch(File.join(foos_dir, "foo_spec.rb"))
-
-        bars_dir = File.join(Dir.tmpdir, "shared/spec/bars")
-        FileUtils.mkdir_p bars_dir
-        FileUtils.touch(File.join(bars_dir, "bar_spec.rb"))
-
-        FileUtils.ln_s bars_dir, File.join(project_dir, "spec/bars")
-
-        FileUtils.cd(project_dir) do
-          expect(loaded_files).to contain_exactly(
-            "spec/bars/bar_spec.rb",
-            "spec/foos/foo_spec.rb"
-          )
-        end
-      end
-
-      it "works on a more complicated example (issue 1113)" do
-        FileUtils.mkdir_p("subtrees/DD/spec")
-        FileUtils.mkdir_p("spec/lib")
-        FileUtils.touch("subtrees/DD/spec/dd_foo_spec.rb")
-        FileUtils.ln_s(File.join(project_dir, "subtrees/DD/spec"), "spec/lib/DD")
-
-        expect(loaded_files).to contain_exactly("spec/lib/DD/dd_foo_spec.rb")
-      end
-    end
+    it_behaves_like "handling symlinked directories when loading spec files"
   end
 end