+++ Changelog.md
@@ -44,6 +44,8 @@ Bug Fixes:
   when it matches no files, by passing along a `--pattern` option to
   the `rspec` command, rather than resolving the file list and passing
   along the files individually. (Evgeny Zislis, #1653)
+* Fix rake task default pattern so that it follows symlinks properly.
+  (Myron Marston, #1672)
 
 ### 3.0.4 / 2014-08-14
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.0.3...v3.0.4)